requirements.txt
# Core Framework
Flask>=2.0.0
flask-cors>=3.0.0

# OpenAI Integration
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.28.0

# Authentication
PyJWT>=2.0.0

# Google OAuth
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# Payment Gateway
razorpay>=1.4.0

# WSGI Server for Production
gunicorn>=20.1.0